# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Simple api for business profiles.",
    "version": "1.0.0",
    "title": "Business API",
    "contact": {
      "email": "atrope@gmail.com"
    }
  },
  "host": "rocky-chamber-96945.herokuapp.com",
  "basePath": "/",
  "tags": [
    {
      "name": "category",
      "description": "Add/List/Update/Remove category"
    },
    {
      "name": "company",
      "description": "Add/List/Update/Remove company"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/category": {
      "get": {
        "tags": [
          "category"
        ],
        "description": "Returns all categories",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "category response",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Category"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "category"
        ],
        "description": "Creates a new Category",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "category",
            "in": "body",
            "description": "Category to add",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewCategory"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "newly create category",
            "schema": {
              "$ref": "#/definitions/Category"
            }
          }
        }
      }
    },
    "/category/{id}": {
      "delete": {
        "tags": [
          "category"
        ],
        "description": "deletes a category - deletes all companies also assigned to it",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of category to delete",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "pet deleted"
          }
        }
      },
      "put": {
        "tags": [
          "category"
        ],
        "description": "Edit a Category",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "category",
            "in": "body",
            "description": "Category to edit",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewCategory"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "ID of category to delete",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "edited category",
            "schema": {
              "$ref": "#/definitions/Category"
            }
          }
        }
      },
      "get": {
        "tags": [
          "category"
        ],
        "description": "Gets a single category",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of category to fetch",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Category",
            "schema": {
              "$ref": "#/definitions/Category"
            }
          }
        }
      }
    },
    "/company": {
      "get": {
        "tags": [
          "company"
        ],
        "description": "Returns all companies",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "category response",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Company"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "company"
        ],
        "description": "Creates a new Category",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "category",
            "in": "body",
            "description": "Company to add",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewCompany"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "newly create Company",
            "schema": {
              "$ref": "#/definitions/Company"
            }
          }
        }
      }
    },
    "/company/{id}": {
      "delete": {
        "tags": [
          "company"
        ],
        "description": "deletes a company - deletes all companies also assigned to it",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of company to delete",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "pet deleted"
          }
        }
      },
      "put": {
        "tags": [
          "company"
        ],
        "description": "Edit a Category",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "company",
            "in": "body",
            "description": "Company to edit",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewCompany"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "ID of category to delete",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "edited company",
            "schema": {
              "$ref": "#/definitions/Company"
            }
          }
        }
      },
      "get": {
        "tags": [
          "company"
        ],
        "description": "Gets a single company",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of Company to fetch",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Company",
            "schema": {
              "$ref": "#/definitions/Company"
            }
          }
        }
      }
    },
    "/company/category/{id}": {
      "get": {
        "tags": [
          "company"
        ],
        "description": "Returns all companies within the category",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of company to delete",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Company",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Company"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Category": {
      "type": "object",
      "allOf": [
        {
          "properties": {
            "_id": {
              "type": "integer"
            },
            "name": {
              "type": "string"
            }
          }
        }
      ]
    },
    "NewCategory": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
    "Company": {
      "type": "object",
      "allOf": [
        {
          "properties": {
            "_id": {
              "type": "integer"
            },
            "name": {
              "type": "string"
            },
            "category": {
              "type": "integer"
            },
            "address": {
              "$ref": "#/definitions/Address"
            }
          }
        }
      ]
    },
    "NewCompany": {
      "type": "object",
      "allOf": [
        {
          "properties": {
            "name": {
              "type": "string"
            },
            "category": {
              "type": "integer"
            },
            "address": {
              "$ref": "#/definitions/Address"
            }
          }
        }
      ]
    },
    "Address": {
      "type": "object",
      "properties": {
        "street": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "state": {
          "type": "string"
        },
        "zip": {
          "type": "integer"
        }
      }
    }
  }
}